% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{class_boolean}
\alias{class_boolean}
\title{Classify cells with boolean logic}
\usage{
class_boolean(obj, classes, drop = TRUE)
}
\arguments{
\item{obj}{Cellpypes object.}

\item{classes}{Character vector with one or more class names.}

\item{drop}{If TRUE (default) and only a single class is supplied,
a boolean vector is returned. A boolean matrix otherwise.}
}
\value{
A boolean vector or matrix.
}
\description{
Classify cells with boolean logic
}
