% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{class_to_deseq2}
\alias{class_to_deseq2}
\title{Create DESeq2 object for a given cell type}
\usage{
class_to_deseq2(obj, meta_df, class, design = ~condition)
}
\arguments{
\item{obj}{A pype object, with rules for at least one cell type.}

\item{meta_df}{Data frame where each column helps to identify a pseudobulk.
Typical columns of meta_df are for example patient, treatment and
cell type -- anything
that uniquely identifies a replicate / batch / 10x run.
Each row in meta_df corresponds to a single cell in your raw count matrix.}

\item{class}{The name of cellpypes class for which you want to test
for differential expression.}

\item{design}{A formula based on columns in \code{meta_df}.
To test differential expression between two groups
in meta_df$condition, use formula \code{~ condition}.
More complex formulas (e.g. with interactions) are possible, for example
\code{~ genotype + treatment + genotype:treatment}.}
}
\value{
A DESeq2 object (e.g. dds)
}
\description{
Create a DESeq2 data set (`dds' in the
\href{www.bioconductor.org/packages/release/bioc/vignettes/DESeq2/inst/doc/DESeq2.html}{DESeq2 vignette})
for the specified class (cell type).
}
\examples{
data("simulated_umis") 
# Meta data
ncells <- ncol(simulated_umis$raw)
dummy_variable <- function(x) factor(sample(x, ncells, replace=TRUE))
meta_data <- data.frame(patient=dummy_variable(paste0("patient", 1:6)),
                        treatment=dummy_variable(c("control", "treated")))
dds <- simulated_umis         \%>\% 
  rule("T", "CD3E",">", 1e-4) \%>\%
  class_to_deseq2(meta_data, "T", ~ treatment)
  
  
}
