% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_pooling.R
\name{evaluate_rule}
\alias{evaluate_rule}
\title{Title}
\usage{
evaluate_rule(obj, class, feature, operator, threshold)
}
\arguments{
\item{obj}{}

\item{class}{}

\item{feature}{}

\item{operator}{}

\item{threshold}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
