% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{feat}
\alias{feat}
\title{Feature plots: Color gene expression in 2D embeddings}
\usage{
feat(obj, features, ...)
}
\arguments{
\item{obj}{Cellpypes object, i.e. a list with four slots: raw, embed, neighbors, totalUMI.}

\item{features}{A vector of genes (features) to colour by.}

\item{...}{Arguments passed to cowplot's \link[cowplot]{plot_grid} function,
for example ncol or rel_widths.}
}
\value{
A ggplot object (assembled by cowplot).
}
\description{
Highlight gene expression in UMAP embeddings, for example.
}
\examples{
feat(simulated_umis, "CD3E")
}
