% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{feat_data}
\alias{feat_data}
\title{Title}
\usage{
feat_data(obj, feature_name)
}
\arguments{
\item{obj}{A cellpypes object, see section \strong{cellpypes Objects} below.}

\item{feature_name}{}
}
\value{

}
\description{
Title
}
