% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_knn.R
\name{find_knn}
\alias{find_knn}
\title{Find approximate k-nearest neighbors}
\usage{
find_knn(featureMatrix, k = 50, n_trees = 50)
}
\arguments{
\item{featureMatrix}{Numeric matrix with features in rows, cells in columns.
Rows could be normalized genes or latent dimensions such as principal
components.}

\item{k}{Number of neighbors to find.}

\item{n_trees}{RccpAnnoy builds a forest  of \code{n_trees} trees.
More trees gives higher precision when querying. Default: 50.}
}
\value{

}
\description{
Find approximate k-nearest neighbors
}
\examples{
  # Imagine we have 30 cells and 100 features:
 fmat <- matrix(rnorm(3000), ncol=30)
 nn <- find_knn(fmat,k=15)
 # nn$idx has 30 rows and 15 columns.
}
