% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{is_classes}
\alias{is_classes}
\title{Check if obj$classes looks as expected.
is_class returns FALSE for example in these cases:
is_classes(NULL)
is_classes(data.frame())
is_classes(data.frame(class=c("T","T"), parent=c("..root..","..root..")))}
\usage{
is_classes(classes)
}
\arguments{
\item{classes}{The obj$classes you want to check.}
}
\value{
logical scalar.
}
\description{
Check if obj$classes looks as expected.
is_class returns FALSE for example in these cases:
is_classes(NULL)
is_classes(data.frame())
is_classes(data.frame(class=c("T","T"), parent=c("..root..","..root..")))
}
