% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{is_rules}
\alias{is_rules}
\title{Check if obj$rules looks as expected.}
\usage{
is_rules(rules)
}
\arguments{
\item{rules}{The obj$rules slot of a cellpypes object.}
}
\value{
logical scalar
}
\description{
Check if obj$rules looks as expected.
}
