% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_classes}
\alias{plot_classes}
\title{Call and visualize 'classify' function}
\usage{
plot_classes(obj, ..., point_size = 0.4, point_size_legend = 2, base_size = 15)
}
\arguments{
\item{obj}{A cellpypes object, see section \strong{cellpypes Objects} below.}

\item{...}{Same parameters as \link[cellpypes]{classify}: classes,
replace_overlap_with, replace_unassigned_with.}

\item{point_size}{Dot size used by \link[ggplot2]{geom_point}.}

\item{point_size_legend}{Dot size displayed in legend.
Legend colors are easier to read with larger points.}

\item{base_size}{The base_size of \link[ggplot2]{theme_bw}, i.e.
how large text is displayed. Default: 15.}
}
\value{

}
\description{
Call and visualize 'classify' function
}
\section{cellpypes Objects}{

A cellpypes object is a \link[base]{list} with four slots:
\itemize{
\item{\code{raw }}{(sparse) matrix with genes in rows, cells in columns}
\item{\code{totalUMI }}{the colSums of obj$raw}
\item{\code{embed }}{two-dimensional embedding of the cells, provided as data.frame
or tibble with two columns and one row per cell.}
\item{\code{neighbors }}{index matrix with one row per cell and k columns, where
k is the number of nearest neighbors (we recommend 15<k<100, e.g. k=50).
Here are two ways to get the neighbors index matrix:
\itemize{
\item Use find_knn(featureMatrix)$idx, where featureMatrix could be principal components, latent variables or normalized genes (features in rows, cells in columns).
\item use as(seurat@graphs[\link{"RNA_nn"}], "dgCMatrix")>.1 to extract the kNN
graph computed on RNA. This also works with RNA_snn, wknn/wsnn or any other
available graph – check with names(seurat@graphs).
}
}
}
}

\examples{
plot_classes(rule(simulated_umis, "T", "CD3E",">", 1))
}
