% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pype_rule_template.R
\name{pype_code_template}
\alias{pype_code_template}
\title{Generate code template for cellpype rules}
\usage{
pype_code_template(n_rules = 3)
}
\arguments{
\item{n_rules}{Number of lines (rules) to generate}
}
\value{
Prints rules to the consoles.
}
\description{
This function \link[cellpypes]{rule} code snippet with neat
text alignment to the console.
Paste this into your script and start changing the rules.
}
\examples{
pype_code_template()
}
