% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pype_from_seurat.R
\name{pype_from_seurat}
\alias{pype_from_seurat}
\title{Convert Seurat to cellpypes object.}
\usage{
pype_from_seurat(seurat)
}
\arguments{
\item{seurat}{A Seurat object.}
}
\value{
A cellpypes object.
}
\description{
Start cellpyping a Seurat object.
This function saves the user from building his own cellpypes object,
which is done with \code{list(umi, neighbors,embed, totalUMI)}.
}
