% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{rule}
\alias{rule}
\title{Add a cell type rule.}
\usage{
rule(
  obj,
  class,
  feature,
  operator = ">",
  threshold,
  parent = NULL,
  use_CP10K = TRUE
)
}
\arguments{
\item{obj}{A cellpypes object, see section \strong{cellpypes Objects} below.}

\item{class}{}

\item{feature}{}

\item{operator}{One of the following: \code{c("<", ">", "<=", ">=")}.}

\item{threshold}{}

\item{parent}{The parent class (e.g. "T" or "T cell").
If NULL, new classes get "..root.." and
existing classes keep their current parent.}

\item{use_CP10K}{Is the threshold provided as counts per 10 thousand UMIs
(CP10K)? If TRUE, cellpypes multiplies the threshold with 1e-4.}
}
\value{

}
\description{
This is the heart of cellpypes and best used in
pipes \code{\%>\%}. Check out examples at
https://github.com/FelixTheStudent/cellpypes!
}
\section{cellpypes Objects}{

A cellpypes object is a \link[base]{list} with four slots:
\itemize{
\item{\code{raw }}{(sparse) matrix with genes in rows, cells in columns}
\item{\code{totalUMI }}{the colSums of obj$raw}
\item{\code{embed }}{two-dimensional embedding of the cells, provided as data.frame
or tibble with two columns and one row per cell.}
\item{\code{neighbors }}{index matrix with one row per cell and k columns, where
k is the number of nearest neighbors (we recommend 15<k<100, e.g. k=50).
Here are two ways to get the neighbors index matrix:
\itemize{
\item Use find_knn(featureMatrix)$idx, where featureMatrix could be principal components, latent variables or normalized genes (features in rows, cells in columns).
\item use as(seurat@graphs[\link{"RNA_nn"}], "dgCMatrix")>.1 to extract the kNN
graph computed on RNA. This also works with RNA_snn, wknn/wsnn or any other
available graph – check with names(seurat@graphs).
}
}
}
}

\examples{
# T cells are CD3E+:
obj <- rule(simulated_umis, "T", "CD3E", ">", .1)
# T cells are MS4A1-:
obj <- rule(obj, "T", "MS4A1", "<", 1)
# Tregs are a subset of T cells:
obj <- rule(obj, "Treg", "FOXP3", ">", .1, parent="T") 

}
