% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_umis.R
\docType{data}
\name{simulated_umis}
\alias{simulated_umis}
\title{Simulated scRNAseq data}
\format{
A list with 4 entries:
\describe{
\item{raw}{Raw (unnormalized) UMI counts for a handful of genes, last row are totalUMI. }
\item{neighbors}{Indices of each cell's 50 nearest neighbors.}
\item{embed}{Simulated UMAP embedding.}
\item{celltype}{Cell type label that I used to simulate the data.}
}
}
\source{
Very simple simulation (c.f. data-raw/simulated_umis.R in source code).
}
\usage{
simulated_umis
}
\description{
This data serves to develop cellpypes and to illustrate its functionality.
I made it up entirely.
}
\keyword{datasets}
