% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{tree_ancestry}
\alias{tree_ancestry}
\title{Find parent, parent's parent and so on for a class using recursive programming}
\usage{
tree_ancestry(classes, class)
}
\arguments{
\item{classes}{The class definitions of a cellpypes object, i.e. obj$classes.}

\item{class}{A character vector of length one with the class.}
}
\value{
Character vector with the ancestry of a class.
}
\description{
Find parent, parent's parent and so on for a class using recursive programming
}
\keyword{internal}
