% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{tree_descendants}
\alias{tree_descendants}
\title{Find child, child's child and so on for class(es) using recursive programming}
\usage{
tree_descendants(classes, class, leafs)
}
\arguments{
\item{classes}{The class definitions of a cellpypes object, i.e. obj$classes.}

\item{class}{A character vector with one or multiple classes.}

\item{leafs}{Has to be the output of tree_leaf_nodes(classes). It's passed
as argument so that tree_leaf_nodes is not executed in each recursion.}
}
\value{
Character vector with the descendants of a class.
}
\description{
Find child, child's child and so on for class(es) using recursive programming
}
\keyword{internal}
