% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{tree_leaf_nodes}
\alias{tree_leaf_nodes}
\title{Finds leaf nodes, i.e. classes without children}
\usage{
tree_leaf_nodes(classes)
}
\arguments{
\item{classes.}{Put obj$classes here, i.e. a tree of class definitions
created with the rule function.}
}
\value{

}
\description{
Finds leaf nodes, i.e. classes without children
}
