% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify cells on previously defined rules}
\usage{
classify(
  obj,
  classes = NULL,
  replace_overlap_with = "Unassigned",
  replace_unassigned_with = "Unassigned",
  return_logical_matrix = FALSE
)
}
\arguments{
\item{obj}{}

\item{classes}{Character vector with one or more class names.
If NULL (the default), all classes for which rules exist are used.}

\item{replace_overlap_with}{}

\item{replace_unassigned_with}{}

\item{return_logical_matrix}{}
}
\value{
I think the order of factor levels is the same as classes, and
if is.null(classes) then it's the same as in obj$classes$class, i.e. the
order in which classes were created by the user. I am not sure, however,
but realize this is important (see e.g. f1_all.Rmd, building an f1-data.frame
relies on it).
}
\description{
Classify cells on previously defined rules
}
