% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify cells with boolean logic}
\usage{
classify(
  obj,
  classes = NULL,
  replace_overlap_with = "Unassigned",
  replace_unassigned_with = "Unassigned",
  return_logical_matrix = FALSE
)
}
\arguments{
\item{obj}{Cellpypes object.}

\item{classes}{Character vector with one or more class names.}

\item{drop}{If TRUE (default) and only a single class is supplied,
a boolean vector is returned. A boolean matrix otherwise.}
}
\value{
A boolean vector or matrix.
}
\description{
Classify cells with boolean logic
}
