% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify cells on previously defined rules}
\usage{
classify(
  obj,
  classes = NULL,
  replace_overlap_with = "Unassigned",
  replace_unassigned_with = "Unassigned",
  return_logical_matrix = FALSE
)
}
\arguments{
\item{obj}{A cellpypes object, see section \strong{cellpypes Objects} below.}

\item{classes}{Character vector with one or more class names.
If NULL (the default), all classes for which rules exist are used.}

\item{replace_overlap_with}{}

\item{replace_unassigned_with}{}

\item{return_logical_matrix}{}
}
\value{
I think the order of factor levels is the same as classes, and
if is.null(classes) then it's the same as in obj$classes$class, i.e. the
order in which classes were created by the user. I am not sure, however,
but realize this is important (see e.g. f1_all.Rmd, building an f1-data.frame
relies on it).
}
\description{
Classify cells on previously defined rules
}
\section{cellpypes Objects}{

A cellpypes object is a \link[base]{list} with four slots:
\itemize{
\item{\code{raw }}{(sparse) matrix with genes in rows, cells in columns}
\item{\code{totalUMI }}{the colSums of obj$raw}
\item{\code{embed }}{two-dimensional embedding of the cells, provided as data.frame
or tibble with two columns and one row per cell.}
\item{\code{neighbors }}{index matrix with one row per cell and k columns, where
k is the number of nearest neighbors (we recommend 15<k<100, e.g. k=50).
Here are two ways to get the neighbors index matrix:
\itemize{
\item Use find_knn(featureMatrix)$idx, where featureMatrix could be principal components, latent variables or normalized genes (features in rows, cells in columns).
\item use as(seurat@graphs[\link{"RNA_nn"}], "dgCMatrix")>.1 to extract the kNN
graph computed on RNA. This also works with RNA_snn, wknn/wsnn or any other
available graph – check with names(seurat@graphs).
}
}
}
}

\examples{
classify(rule(simulated_umis, "Tcell", "CD3E", ">", 1))
}
