% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{feat}
\alias{feat}
\title{Feature plots: Color gene expression in 2D embeddings}
\usage{
feat(obj, features, ...)
}
\arguments{
\item{obj}{}

\item{features}{}

\item{...}{}
}
\value{

}
\description{
Feature plots: Color gene expression in 2D embeddings
}
