% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_classes}
\alias{plot_classes}
\title{Call and visualize 'classify' function}
\usage{
plot_classes(obj, ...)
}
\arguments{
\item{obj}{}

\item{...}{Same parameters as \link[cellpypes]{classify}.}
}
\value{

}
\description{
Call and visualize 'classify' function
}
