% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_last}
\alias{plot_last}
\title{Plot the last modified rule or class}
\usage{
plot_last(obj, show_feat = TRUE, what = "rule")
}
\arguments{
\item{obj}{A cellpypes object.}

\item{show_feat}{If TRUE (default), a second panel shows the feature plot of
the relevant gene.}

\item{what}{Either "rule" or "class".}
}
\value{
The plot is drawn as side-effect with \code{print}. The function
returns obj invisibly, so that you can use it in a pipe.
}
\description{
Plot the last modified rule or class
}
