% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_last}
\alias{plot_last}
\title{Plot the last modified rule or class}
\usage{
plot_last(obj, show_feat = TRUE, what = "rule", legend_rel_width = 0.3)
}
\arguments{
\item{obj}{A cellpypes object.}

\item{show_feat}{If TRUE (default), a second panel shows the feature plot of
the relevant gene.}

\item{what}{Either "rule" or "class".}
}
\value{
Returns a ggplot2 object with the plot.
}
\description{
Plot the last modified rule or class
}
\examples{
plot_last(rule(simulated_umis, "T", "CD3E",">", 1))
}
