% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_pooling.R
\name{pool_across_neighbors}
\alias{pool_across_neighbors}
\title{Title}
\usage{
pool_across_neighbors(x, neighbors)
}
\arguments{
\item{x}{}

\item{neighbors}{}
}
\value{

}
\description{
Title
}
\examples{

}
