% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{rule}
\alias{rule}
\title{Add}
\usage{
rule(obj, class, feature, operator = ">", threshold, parent = NULL)
}
\arguments{
\item{obj}{}

\item{class}{}

\item{feature}{}

\item{operator}{}

\item{parent}{}

\item{t}{}
}
\value{

}
\description{
Add
}
