% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{rule}
\alias{rule}
\title{Add a cell type rule.}
\usage{
rule(
  obj,
  class,
  feature,
  operator = ">",
  threshold,
  parent = NULL,
  use_CP10K = TRUE
)
}
\arguments{
\item{obj}{}

\item{class}{}

\item{feature}{}

\item{operator}{One of the following: \code{c("<", ">", "<=", ">=")}.}

\item{threshold}{}

\item{parent}{The parent class (e.g. "T" or "T cell").
If NULL, new classes get "..root.." and
existing classes keep their current parent.}

\item{use_CP10K}{Is the threshold provided as counts per 10 thousand UMIs
(CP10K)? If TRUE, cellpypes multiplies the threshold with 1e-4.}
}
\value{

}
\description{
Add a cell type rule.
}
