% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{tree_leaf_nodes}
\alias{tree_leaf_nodes}
\title{Finds leaf nodes, i.e. classes without children}
\usage{
tree_leaf_nodes(classes, invert = FALSE)
}
\arguments{
\item{classes}{Put obj$classes here, i.e. a tree of class definitions
created with the rule function.}

\item{invert}{If TRUE, return classes that are NOT leafs instead of leaf nodes.}
}
\value{
Character vector with the tree's leafs.
}
\description{
Finds leaf nodes, i.e. classes without children
}
\keyword{internal}
